% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossdist.R
\name{crossdist}
\alias{crossdist}
\title{Cross-distance matrix calculation}
\usage{
crossdist(m1, m2)
}
\arguments{
\item{m1}{Numeric matrix.}

\item{m2}{Numeric matrix.}
}
\value{
Numeric matrix of distances.
}
\description{
Computes the pairwise Euclidean distance between rows of two matrices.
}
\examples{
mat1 <- matrix(1:4, ncol = 2)
mat2 <- matrix(5:8, ncol = 2)
distance_matrix <- crossdist(mat1, mat2)
}
