% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossdist2.R
\name{crossdist2}
\alias{crossdist2}
\title{Calculate Cross-Distance Matrix}
\arguments{
\item{m1}{A numeric matrix.}

\item{m2}{Another numeric matrix.}
}
\value{
A numeric matrix where the entry at \verb{[i, j]} is the Euclidean distance between row i of \code{m1} and row j of \code{m2}.
}
\description{
This function computes the pairwise Euclidean distance between rows of two matrices.
}
\examples{
mat1 <- matrix(1:4, ncol = 2)
mat2 <- matrix(5:8, ncol = 2)
crossdist2(mat1, mat2)
}
