% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCell.R
\name{FindCell}
\alias{FindCell}
\title{Find cells indicating shared biological features across conditions}
\usage{
FindCell(seuratobj, seuratlist, fullcluster, distmat, firstn = 15)
}
\arguments{
\item{seuratobj}{The Seurat object that all samples/subjects were merged together.}

\item{seuratlist}{A list of Seurat objects, usually can be got by SplitObject().}

\item{fullcluster}{A list of clusters that generated by the function GetCluster().}

\item{distmat}{A list of distance vectors generated by the function FindNNDist().}

\item{firstn}{Positive integer giving the number of nearest neighbors from
each sampled cell to include. It must not exceed the \code{meaningn} value used
to create \code{distmat}. The published method uses 15.}
}
\value{
A list of two vectors: one is for which cluster of which sample will be highlighted and the second one is which cells will be selected.
}
\description{
Find cells indicating shared biological features across conditions
}
\examples{
data(sim_data_sce)
data(sim_result)
sim_data <- SCEtoSeurat(sim_data_sce)
seuratlist <- Seurat::SplitObject(sim_data, split.by = "Study")
result <- FindCell(
  sim_data, seuratlist, sim_result[[1]], sim_result[[3]], firstn = 15
)
head(result[[2]])
}
