% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCEtoSeurat.R
\name{SCEtoSeurat}
\alias{SCEtoSeurat}
\title{Input and Split SingleCellExperiment Data}
\usage{
SCEtoSeurat(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}
}
\value{
A Seurat objects.
}
\description{
This function takes a SingleCellExperiment object and a variable by which to split it,
converts it to a Seurat object, and then splits it according to the specified variable.
}
\examples{
data(sim_data_sce)
seuratobj <- SCEtoSeurat(sim_data_sce)
seuratobj
}
