% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{My Example Dataset}
\format{
An example PBMC data with Seurat format
\describe{
\item{assays}{Count matrix}
\item{meta.data}{Meta data for each cell}
}
}
\usage{
sim_data
}
\description{
A short description of your dataset.
}
\keyword{datasets}
