% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummCluster.R
\name{SummCluster}
\alias{SummCluster}
\title{Get maximum number of broad clusters}
\usage{
SummCluster(fullcluster)
}
\arguments{
\item{fullcluster}{A list of clusters that generated by the function GetCluster()}
}
\value{
A list
}
\description{
Get maximum number of broad clusters
}
\examples{
data(sim_result)
cluster_summary <- SummCluster(sim_result[[1]])
cluster_summary$nbroad
}
