% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcuSCIR.R
\name{CalcuSCIR}
\alias{CalcuSCIR}
\title{Calculate SCIntRuler}
\usage{
CalcuSCIR(fullcluster, seuratlist, testres, p = 0.1)
}
\arguments{
\item{fullcluster}{A list of clusters that generated by the function \code{GetCluster()}}

\item{seuratlist}{A list of Seurat objects, usually can be got by \code{SplitObject()}.}

\item{testres}{Result from function \code{PermTest()}}

\item{p}{P-value that will be used as the cut-off, default value is 0.1}
}
\value{
SCIntRuler
}
\description{
Calculate SCIntRuler
}
\examples{
data(sim_result)
data(sim_data)
# Create example data for fullcluster (mock data)
seuratlist <- SplitObject(sim_data, split.by = "Study")
# fullcluster <- GetCluster(seuratlist)
# testres <- PermTest(fullcluster,distmat,15)
CalcuSCIR(sim_result[[1]], seuratlist, sim_result[[4]])
}
