% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCell.R
\name{FindCell}
\alias{FindCell}
\title{Find cells indicating shared biological features across conditions}
\usage{
FindCell(seuratobj, seuratlist, fullcluster, distmat, firstn = 15)
}
\arguments{
\item{seuratobj}{The \verb{Seurat object} that all samples/subjects were merged together.}

\item{seuratlist}{A list of Seurat objects, usually can be got by \code{SplitObject()}.}

\item{fullcluster}{A list of clusters that generated by the function \code{GetCluster()}.}

\item{distmat}{A list of distance vectors generated by the function \code{FindNNDist()}.}

\item{firstn}{The number of nearest cells were detected that you want to include in the permutation test. Default to be 15.}
}
\value{
A list of two vectors: one is for which cluster of which sample will be highlighted and the second one is which cells will be selected.
}
\description{
Find cells indicating shared biological features across conditions
}
\examples{
\dontrun{
# Create example Seurat object
seuratobj <- CreateSeuratObject(...)
# Create example list of Seurat objects
seuratlistseuratlist <- SplitObject(seuratobj, split.by = "Study")
# Create example fullcluster (mock data)
fullcluster <- GetCluster(seuratlist)
# Create example distmat (mock data)
distmat <- FindNNDist(fullcluster, normCount, meaningn = 20)

FindCell(seuratobj, seuratlist, fullcluster, distmat, 15)
}
}
