% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNNDist.R
\name{FindNNDist}
\alias{FindNNDist}
\title{Find the nearest neighbors}
\usage{
FindNNDist(fullcluster, normCount, meaningn = 20)
}
\arguments{
\item{fullcluster}{A list of clusters that generated by the function \code{GetCluster}.}

\item{normCount}{A list of normalized gene count matrix generated by the function \code{NormData}.}

\item{meaningn}{default to be 20}
}
\value{
A list of distance vectors
}
\description{
Find the nearest neighbors
}
\examples{
\dontrun{
# Create example data for fullcluster (mock data)
fullcluster <- GetCluster(seuratlist)
# Create example data for normCount (mock data)
normCount <- NormData(seuratlist)
# Define meaningn
meaningn <- 20

FindNNDist(fullcluster, normCount, meaningn = meaningn)
}
}
