% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNNDistC.R
\name{FindNNDistC}
\alias{FindNNDistC}
\title{Find the nearest neighbors}
\usage{
FindNNDistC(fullcluster, normCount, meaningn = 20)
}
\arguments{
\item{fullcluster}{A list of clusters that generated by the function GetCluster().}

\item{normCount}{A list of normalized gene count matrix generated by the function NormData().}

\item{meaningn}{default to be 20}
}
\value{
A list of distance vectors
}
\description{
Find the nearest neighbors
}
\examples{
data(sim_result)
meaningn <- 20
FindNNDistC(sim_result[[1]], sim_result[[2]], meaningn = meaningn)
}
