% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormData.R
\name{NormData}
\alias{NormData}
\title{Normalized RNA data matrix}
\usage{
NormData(seuratlist)
}
\arguments{
\item{seuratlist}{A list of Seurat objects, usually can be got by SplitObject().}
}
\value{
A list of matrix.
}
\description{
Normalized RNA data matrix
}
\examples{
data(sim_data_sce)
sim_data <- SCEtoSeurat(sim_data_sce)
seuratlist <- Seurat::SplitObject(sim_data, split.by = "Study")
normCount <- NormData(seuratlist)
}
