% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormData.R
\name{NormData}
\alias{NormData}
\title{Normalized RNA data matrix}
\usage{
NormData(seuratlist)
}
\arguments{
\item{seuratlist}{A list of Seurat objects, usually can be got by \code{SplitObject()}.}
}
\value{
A list of matrix.
}
\description{
Normalized RNA data matrix
}
\examples{
\dontrun{
Assuming 'seurat_object_list' is a list of Seurat objects
seuratlist <- SplitObject(seuratobj, split.by = "Study")
normCount <- NormData(seuratlist)
}
}
