% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermTest.R
\name{PermTest}
\alias{PermTest}
\title{Permutation Test}
\usage{
PermTest(fullcluster, distmat, firstn)
}
\arguments{
\item{fullcluster}{A list of clusters that generated by the function GetCluster()}

\item{distmat}{A list of distance vectors generated by the function FindNNDist().}

\item{firstn}{The number of nearest cells were detected that you want to include in the permutation test.}
}
\value{
A list of two lists, one is the relative within-between distance and another is p-value of permutation test. Default to be 15.
}
\description{
Permutation Test
}
\examples{
data(sim_result)
testres <- PermTest(sim_result[[1]], sim_result[[3]],15)
}
