% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermTest.R
\name{PermTest}
\alias{PermTest}
\title{Permutation Test}
\usage{
PermTest(fullcluster, distmat, firstn = 15)
}
\arguments{
\item{fullcluster}{A list of clusters that generated by the function GetCluster()}

\item{distmat}{A list of distance vectors generated by the function FindNNDist().}

\item{firstn}{Positive integer giving the number of nearest neighbors from
each sampled cell to include. It must not exceed the \code{meaningn} value used
to create \code{distmat}. The published method uses 15.}
}
\value{
A list containing the relative within-between distance matrix
(\code{allrevDiff}) and permutation-test p-values (\code{allP}).
}
\description{
Permutation Test
}
\examples{
data(sim_result)
# fullcluster <- GetCluster(seuratlist)
# Assuming 'distmat' is a list of normalized gene count matrix
# distmat <- FindNNDist(fullcluster, normCount, meaningn = 20)
testres <- PermTest(sim_result[[1]], sim_result[[3]], 15)
}
