% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSCIR.R
\name{PlotSCIR}
\alias{PlotSCIR}
\title{Plot SCIntRuler}
\usage{
PlotSCIR(fullcluster, seuratlist, testres, legendtitle = NULL, title = NULL)
}
\arguments{
\item{fullcluster}{A list of clusters that generated by the function \code{GetCluster}.}

\item{seuratlist}{A list of Seurat objects, usually can be got by \code{SplitObject()}.}

\item{testres}{Result from function \code{PermTest()}}

\item{legendtitle}{Title of legend, default to be \code{NULL}}

\item{title}{Title of figure, default to be \code{NULL}}
}
\value{
A ggplot2 object
}
\description{
Plot SCIntRuler
}
\examples{
# Create example data for fullcluster (mock data)
data(sim_data)
data(sim_result)
seuratlist <- SplitObject(seuratobj, split.by = "Study")
# fullcluster <- GetCluster(seurat_object_list)
# testres <- PermTest(fullcluster,distmat,15)
PlotSCIR(sim_result[[1]], seuratlist, sim_result[[4]])
}
