% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_result_data.R
\docType{data}
\name{sim_result}
\alias{sim_result}
\title{My Example Dataset}
\format{
An result example data
\describe{
\item{fullcluster}{A runable example of GetCluster, which is a list of clusters for each study.}
\item{normCount}{A runable example of NormData, which is a list of normalized RNA expression matrixs for each study.}
\item{distmat}{A runable example of FindNNDist, which is a list of distance matrixs for each study.}
\item{testres}{A runable example of CalcuSCIR, which is a list of test results for each study.}
}
}
\usage{
sim_result
}
\value{
Simulation data to examplify the usage of the method.
}
\description{
An result example data with results from different functions.
}
\examples{
# Load the data
data("sim_result")
}
\keyword{datasets}
