% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_result_data.R
\docType{data}
\name{sim_result}
\alias{sim_result}
\title{My Example Dataset}
\format{
An result example data
\describe{
\item{sim_result[\link{1}]}{fullcluster}
\item{sim_result[\link{2}]}{normCount}
\item{sim_result[\link{3}]}{distmat}
}
}
\usage{
sim_result
}
\description{
A short description of your dataset.
}
\keyword{datasets}
