% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_result_data.R
\docType{data}
\name{sim_result}
\alias{sim_result}
\title{My Example Dataset}
\format{
An result example data
\describe{
\item{\code{fullcluster}}{A runable example of fullcluster, which is a list of clusters for each study.}
\item{\code{normCount}}{A runable example of normCount, which is a list of normalized RNA expression matrixs for each study.}
\item{\code{distmat}}{A runable example of distmat, which is a list of distance matrixs for each study.}
\item{\code{testres}}{A runable example of testres,  which is a list of test results for each study.}
}
}
\usage{
sim_result
}
\value{
Simulation data to examplify the usage of the method.
}
\description{
A short description of your dataset.
}
\examples{
# Load the data
data("sim_result")
}
\keyword{datasets}
